Newspaperstate and Better Unittests (slow)

+++ conflict_interface/data_types/player_state/player_profile.py
@@ -68,7 +68,7 @@ class PlayerProfile(GameObject):
     title: str
     nation_name: str
     nation_adjective: str
-    average_national_morale: float
+    average_national_morale: int
     computer_player: bool
     native_computer: bool
     site_user_id: int
